refactor: the web ui codebase
