--- src/web_ui/src/components/Forms/Signup/Signup.css
@@ -1,17 +1,0 @@
-.data-form-inputs {
-    width: 75%;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  
-  input {
-    width: 100%;
-    padding: 6px 20px;
-    margin: 8px 0;
-    box-sizing: border-box;
-  }
-  
-  form {
-    padding-bottom: 5%;
-  }
