+++ src/web_ui/src/components/Layout/Footer/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import Copyright from '../../common/Copyright';
 import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
+import Copyright from '../../common/Copyright';
 
 const useStyles = makeStyles((theme) => ({
   footer: {
     backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(6),
+    padding: theme.spacing(6)
   },
 }));
 
@@ -14,14 +14,14 @@ export default function Footer() {
   const classes = useStyles();
 
   return (
-      <footer className={classes.footer}>
-        <Typography variant="h6" align="center" gutterBottom>
-          Footer
-        </Typography>
-        <Typography variant="subtitle1" align="center" color="textSecondary" component="p">
-          Something here to give the footer a purpose!
-        </Typography>
-        <Copyright />
+    <footer className={classes.footer}>
+      <Typography variant="h6" align="center" gutterBottom>
+        Footer
+      </Typography>
+      <Typography variant="subtitle1" align="center" color="textSecondary" component="p">
+        Something here to give the footer a purpose!
+      </Typography>
+      <Copyright />
     </footer>
   );
 }
