+++ src/web_ui/src/components/Pages/Homepage.js
@@ -1,6 +1,4 @@
 import React from 'react';
-import ProductSearch from '../common/ProductSearch';
-import Layout from '../Layout';
 import Button from '@material-ui/core/Button';
 import Card from '@material-ui/core/Card';
 import CardActions from '@material-ui/core/CardActions';
@@ -12,6 +10,8 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import { List, ListItem } from '@material-ui/core';
+import ProductSearch from '../common/ProductSearch';
+import Layout from '../Layout';
 
 const useStyles = makeStyles((theme) => ({
   flex: {
@@ -54,68 +54,65 @@ export default function Album() {
   const classes = useStyles();
 
   return (
-    <React.Fragment>
-      <Layout>
-        {/* Hero unit */}
-        <div className={classes.heroContent}>
-          <Container maxWidth="sm">
-            <ProductSearch />
-          </Container>
-        </div>
-        <Container className={classes.cardGrid} maxWidth="md">
-          {/* End hero unit */}
-          <Grid container spacing={4}>
-            {cards.map((card) => (
-              <Grid item key={card} xs={12} sm={6} md={4}>
-                <Card className={classes.card}>
-                  <CardMedia
-                    className={classes.cardMedia}
-                    image="https://source.unsplash.com/random"
-                    title="Vegetables"
-                  />
-                  <CardContent className={classes.cardContent}>
-                    <Grid container spacing={4}>
-                      <Grid item xs={6}>
-                        <Typography color="textSecondary" gutterBottom component="h5">
-                          VEGETABLES
-                        </Typography>
-                      </Grid>
-                      <Grid item xs={6}>
-                        <List className={classes.cardList}>
-                          <ListItem className={classes.cardList} disableGutters>
-                            <Timer className={classes.space} /> 
-                            2 hours ago
-                          </ListItem>
-                        </List>
-                      </Grid>
-                    </Grid>
-                    <Grid container className={classes.spaceTop}>
-                      <Grid item xs={2}>
-                        <LocalOffer fontSize="large"/>
-                      </Grid>
-                      <Grid item xs={3}>
-                        <Typography component="h5">
-                          Dodo
-                        </Typography>
-                        Fastfarm
-                      </Grid>
-                    </Grid>
-                  </CardContent>
-                  <CardActions>
-                    <Button size="small" className={classes.flex}>
-                      VIEW PRODUCT
-                    </Button>
-                    <Button size="small" className={classes.flex}>
-                      VIEW FARM
-                    </Button>
-                  </CardActions>
-                </Card>
-              </Grid>
-            ))}
-          </Grid>
+    <Layout>
+      {/* Hero unit */}
+      <div className={classes.heroContent}>
+        <Container maxWidth="sm">
+          <ProductSearch />
         </Container>
-      </Layout>
-      {/* End footer */}
-    </React.Fragment>
+      </div>
+      <Container className={classes.cardGrid} maxWidth="md">
+        {/* End hero unit */}
+        <Grid container spacing={4}>
+          {cards.map((card) => (
+            <Grid item key={card} xs={12} sm={6} md={4}>
+              <Card className={classes.card}>
+                <CardMedia
+                  className={classes.cardMedia}
+                  image="https://source.unsplash.com/random"
+                  title="Vegetables"
+                />
+                <CardContent className={classes.cardContent}>
+                  <Grid container spacing={4}>
+                    <Grid item xs={6}>
+                      <Typography color="textSecondary" gutterBottom component="h5">
+                        VEGETABLES
+                      </Typography>
+                    </Grid>
+                    <Grid item xs={6}>
+                      <List className={classes.cardList}>
+                        <ListItem className={classes.cardList} disableGutters>
+                          <Timer className={classes.space} /> 
+                          2 hours ago
+                        </ListItem>
+                      </List>
+                    </Grid>
+                  </Grid>
+                  <Grid container className={classes.spaceTop}>
+                    <Grid item xs={2}>
+                      <LocalOffer fontSize="large" />
+                    </Grid>
+                    <Grid item xs={3}>
+                      <Typography component="h5">
+                        Dodo
+                      </Typography>
+                      Fastfarm
+                    </Grid>
+                  </Grid>
+                </CardContent>
+                <CardActions>
+                  <Button size="small" className={classes.flex}>
+                    VIEW PRODUCT
+                  </Button>
+                  <Button size="small" className={classes.flex}>
+                    VIEW FARM
+                  </Button>
+                </CardActions>
+              </Card>
+            </Grid>
+          ))}
+        </Grid>
+      </Container>
+    </Layout>
   );
 }
